requirements.txt
Flask==2.3.3
flask-cors==4.0.0
flask-limiter==3.5.1
google-generativeai==0.3.2
pinecone-client==2.2.4